:recycle: Refactor draw method and change color in inactive session
