+++ .gitignore
@@ -18,3 +18,4 @@ tmp/
 .windsurfrules
 
 examples/*.csv
+.aider*
